requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.4
Flask-Cors==4.0.0
numpy==1.26.4
shapely==2.1.1
pytest==7.4.2
networkx==3.1
python-socketio==5.8.0
requests==2.31.0
pygame==2.5.2 